feat(@crates/eur-vision): alter capture region signature to make reusable
